requirements.txt
langchain_community==0.0.31
langchain_openai==0.1.1
langchain_text_splitters==0.0.1
pandas==2.0.3
python-dotenv==1.0.1
weaviate_client==4.5.4
